refactor: Split bounty-detail-sidebar-cta.tsx (#209)
- Extracted MobileCTA into its own file
- Extracted shared state into useBountyCTAState hook
- Updated bounty-detail-client.tsx to import MobileCTA from new file
