+++ components/bounty-detail/bounty-detail-client.tsx
@@ -4,7 +4,8 @@ import { useState, useCallback } from "react";
 import { useRouter } from "next/navigation";
 import { AlertCircle, ArrowLeft } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { MobileCTA, SidebarCTA } from "./bounty-detail-sidebar-cta";
+import { SidebarCTA } from "./bounty-detail-sidebar-cta";
+import { MobileCTA } from "./bounty-detail-mobile-cta";
 import { HeaderCard } from "./bounty-detail-header-card";
 import { DescriptionCard } from "./bounty-detail-description-card";
 import { BountyDetailSubmissionsCard } from "./bounty-detail-submissions-card";
